+++ src/main/java/com/chess/moveorder/MoveOrderService.java
@@ -6,7 +6,7 @@ import java.util.Arrays;
 import java.util.List;
 
 public interface MoveOrderService {
-    int getOrder(Move move);
+    double getOrder(Move move);
 
 
 }
