+++ src/main/java/com/chess/moveorder/Takes.java
@@ -8,7 +8,7 @@ import com.chess.Square;
 public class Takes implements MoveOrderService {
 
     @Override
-    public int getOrder(Move move) {
+    public double getOrder(Move move) {
         Position position = move.getPosition();
 
         Square endingSquare = move.getEndingSquare();
